Update mob_groups.sql
This is my first attempt at a pull request. I'm sorry if I'm doing it wrong. (Also in the future i'll edit directly on github and not copy paste from my personal file.) In any case I went through and edited all enemies added with grounds of valor to ensure they matched the wiki. I thought this would be a simple starting point and would allow players to better use the older areas for reaching 99. (I figured if the enemies are already added and in the 81-84 range may as well make them accurate.) I made a few estimations as marked by the text. Also I didn't touch updates to CoP, ToAU and WoTG since much of that is not fully implemented yet.  Though, it shouldn't be too much trouble if that is something worth doing. Any comments appreciated, thanks.