Add capability to log GM command use to database.
Functions as a minimum to bother logging, tier zero commands always ignored. Value of 2=log t2 and above but ignore t1.

Another approach would have been to add a new field to the cmdprops table however this is re-read each and every time a command is used and would add more overhead reading those in addition to requiring an alteration of every existing GM command script. Doing this entirely core side is the most efficient solution unelss you want to rewrite the commandhandler source. Which we probably could. Changing the table to be a mask would be a lot more efficient as well but far less human readable.