CORE: Add trigger for starting Ally Assist
Allies will assist players in battle. Add a generic hook for
starting this behavior that is accesible in lua, so that is can
be scripted to happen at the appropiate point in the instance
script.

Two modes are provided:

ALLY_ASSIST_PLAYER: Attempt to find a player with a battle target,
and assist that player. This is the default mode.

ALLY_ASSIST_RANDOM: Pick a random mob to fight. Some allies do
this in retail.