CORE: Port AllyAssist Routine from Instance -> BCNM
Port a simplified version of the AllyAssist routine I made
for Instances to BCNMs. Makes things easy and works with
the ally if a mob is marked with player allegience.

Unfortunately the two systems do not use the same data structures
so I couldn't just share the code without refactoring or costly
runtime translation.