SCRIPTS/SQL: TOAU-29 Puppet in Peril
BCNM Fight in the Jade Sepulcher

Note: There are minor glitches in the cutscenes because the
NPCs need to be capped from retail. We had no NPCs in this
zone yet. I dumped the NPCs from POLUtils and used the look
from other captured versions where available. For the most
part this worked but a couple of NPCs have minor glitches.

Sources:

http://ffxiclopedia.wikia.com/wiki/Puppet_in_Peril
https://www.bg-wiki.com/bg/Aht_Urhgan_Mission_29
gedads data: http://pastebin.com/w0TqRPMy