CORE: Hack around a crash on Local Vars
If a mob dies during the process of setting a local var, a crash
will occurr in the STL in std::map. I was not able to figure out
why this is happened (will file issue w/ bt), so I suppressed the
sets if the mob is dead.

This could happen frequently in mission battles where mob vars are
broadcast as signal flags to help coordinate the correct behavior