CORE: Add fallback hierachy for onEventFinish / onEventUpdate
Unroll the nested load loop so that you can check if the function
is defined as well as if the file exists. Previously, you would
fall back to Zone.lua if the event NPC did not have a lua file, but
not if it had one but did not define this method.

Also added support for checking the instance scripts as part of the
hierarchy, allowing instances to optionally override the Zone.lua
event handling if there are special needs for the given instance,
or if the author wants to co-locate instance specific CS handling
with the instance script itself for easier maintenance, readability,
and testing since the instance scripts don't get cached.

Hierarchy order: m_event -> Instance (if applicable) -> Zone