CORE: BCNM Mobs that aggro should always aggro
This didn't used to be a problem because of level cap, but now
that many BCNMs have become uncapped on retail, it is possible
for a player to be of high enough level that they don't aggro
anymore.

I have seen many videos of people soloing tough missions at 90-99
on Youtube, and they always still get aggroed.

Thus, if a BCNM mob has the aggro flag set, we'll assume it should
always aggro