CORE/SCRIPTS: Fix Running Away in BCNMs
Running away properly requires a specific sequence of updates to
the run away event in order to finish the cutscene and not get
stuck.

Ironically, someone had documented the right sequence in bcnm.lua,
it just was not implemented. Used these notes, implemented it,
and tested it on TOAU-22 Shields of Diplomacy BCNM.

Also had to colocate delPlayerFromBCNM handling on event finish
with the other ones... that makes this dependent on my TOAU branch
since I refactored that.