Adding in and adjusting missing or wrong item mods.
Update: Fixed the type of pet per the recent submission:
    All = 0,
    Avatar = 1,
    Wyvern = 2,
    Automaton = 3

Also fixed spacing and comments to the comments showed if it was pet, avatar, wyvern, or automaton.

Fix: Well, shit, didn't realize that the upgrade from Evoker's Doublet to +1 causes you to lose, well, all pet mods. Removed the +1 completely.