SQL prep for implementing grouping (shared drop slots) of mob drops. New columns are not used yet.
Planned use: roll dice for each item in the group, highest roll wins. Least rare item wins in the event of a tie. This allows the item droprate (itemRate) to be used as it always has.
The groupRate column might be removed later..It exists for an edge-case situation where there isn't a guaranteed drop in the grouping in the info I have. Easier for me to mass remove it later than strip it out of my branch right now for this commit and then wind up re-adding it if I use it..Even heavily customized servers shouldn't have much trouble with this change.