Fix Conflict in keyitems.lua (TONBERRY_KEY)
There are 2 variables in this file that are exactly the same (TONBERRY_KEY) but refer to different key items with the same name.
This is causing a conflict which prevents the quest "Everyone's Grudge" from allowing players to reset their hate. Since the key items themselves have exactly the same name I just changed the one required for this quest to indicate it's used for the PRIEST for lack of a better way.