Fix(tonberry_dismayer): Update Tonberry Kinq spawn chance
Teo pointed out that there are no accounts of a true spawn chance, and two accounts of a 6-7 hour spawn time. Therefore, I've increased the cooldown to 6 hours and bumped the spawn chance to 20%. Since there are two placeholders, you should see a pop within 6 - 7 hours of the last kill.