trait(daken): Add +100 RAcc bonus to Daken proc
I've added a new int8 accBonus parameter to the GetRangedHitRate function. Since there is no effect on the battle entity itself that says it is doing a Daken throw we cannot use GetRangedAccuracyBonuses like we do for Barrage. I've set it to int8 in case we ever need a penalty in the future.

See here: https://www.bg-wiki.com/bg/Daken