Goblin Wolfman not Spawning at Scrawled Writing
Goblin Wolfman will spawn in his default spawn pos, then move to the Scrawled Writing resulting in linking everything in it's path, losing claim and eventually despawning.

Ideally it's best to use setSpawn to spawn at the onTrade location instead of spawning it else where and using setPos to relocate it.